function/requirements.txt
azure-functions>=1.16.0
requests>=2.31.0
azure-storage-blob>=12.19.0
